+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Auto-Start feature
 
+### Changed
+
+- Set window size on startup using an alternative method.
+
 ## [1.1.0] - 2022-01-26
 
 ### Added